Coded in, and finished dinucleotide analysis to see if doi:10.1038/mt.2012.29 holds across the genome.
Also modified the Gene class such that protein expression dict, ontology dict, and RNA half lives dict are all used to populate class variables on class instantiation as opposed to doing it in subsequent method calls. I don't know if this is actually faster, but it simplifies the code in Analysis.py.